requirements.txt
spotipy>=2.19.0
python-dotenv>=0.19.0
pandas>=1.3.0
requests>=2.25.0
